+++ pom.xml
@@ -20,7 +20,9 @@
     </licenses>
 
     <modules>
+        <module>sso-kit-core</module>
         <module>sso-kit-starter</module>
+        <module>sso-kit-starter-hilla</module>
         <module>sso-kit-keycloak</module>
     </modules>
