+++ sso-kit-core/.gitignore
@@ -1,0 +1,9 @@
+/target/
+.idea/
+.vscode/
+.settings
+.project
+.classpath
+
+*.iml
+.DS_Store
