+++ sso-kit-core/src/main/resources/sso-kit.properties
@@ -1,0 +1,1 @@
+sso-kit.version=${project.version}
