+++ sso-kit-starter-hilla/src/main/java/dev/hilla/sso/starter/SingleSignOnData.java
