+++ sso-kit-starter-hilla/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+dev.hilla.sso.starter.SingleSignOnConfiguration
