+++ sso-kit-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,1 @@
-com.vaadin.sso.starter.SingleSignOnDefaultBeans
 com.vaadin.sso.starter.SingleSignOnConfiguration
